Fix a potential bug in the data_generator module
It seems that sum(class_p) can either return a value of type 'int' or
type 'np.int64' depending on some side effect. The former will result in
an error being raised, because numpy's element-wise division will not be
invoked. To resolve this, class_p is wrapped around a numpy array before
carrying out the division.