Add support for class-wise probability thresholds
The idea is that, when evaluating the validation set, the user can pass
a '--thresholds' flag indicating that optimal thresholds should be
recorded for each class. 'Optimal' means that it is the threshold that
corresponds to the highest F1 score based on a precision-recall curve.
When evaluating the test set, the user can then use these saved
thresholds instead of a hard-coded value. A threshold value of -1 for
the appropriate config variable can be used to achieve this.