Modify logmel feature extraction
The n_overlap configuration has been replaced with hop_length, which is
equal to n_window - n_overlap. Instead of using scipy's spectrogram()
function, the stft() function from librosa is now used. The type of STFT
window to be used is no longer specified (default is Hann) as it is an
insignificant detail. The logmel features are now given in decibels from
full scale (dBFS). This is so that, when the features are loaded, they
be clipped to an arbitrary threshold (currently -90 dBFS).